show-resolved shows UUID, can edit resolved tasks
Modified the show-resolved display to also include the first 8 digits of
the UUID in the first column.

Additionally added the ability to edit resolved tasks by using the uuid:
modifier with the edit command.

CMD_EDIT will check if the command line includes a UUID and load
resolved tasks if so. It will then check to see if it can match the user
entered UUID with any resolved tasks. If only one matches, it is edited.
User is prompted if 0 or more than 1 UUID matches. In the case that more
than 1 UUID matches, the matching UUIDs are listed.