Added uuid: completions. MustGetFromUUID now returns (task, error)
Updated VALID_STATUS_TRANSITIONS to include RESOLVED -> TEMPLATE &&
RESOLVED -> PENDING transistions.

Since I've added completions, there may not not be a need to match a
partial UUID entered by the user to the UUIDs of resolved tasks. I've
left the **incomplete** code commented in taskset.go as a TODO.

Still need to implement some tests and add the functionality to
CMD_MODIFY and CMD_NOTES.