Bring sorting options into NewTaskSet functional options.
We provide a default sorting strategy of P1 first, and then newest
first, if no options are passed. Sorting is still performed by methods
on TaskSet, but these methods are now private.

The order that SortBy options are passed matters, as it does in SQL.

Sorting happens to our TaskSet after tasks are loaded from disk, but
before filtering happens. So for now, our procedure for building a
TaskSet is LOAD -> SORT -> FILTER. Then we can render.