Update Makefile
This is a cold rewrite of the Makefile. Possibly not useful, because I might have not completly understood all the necessary steps or intentions. The new Makefile merges functionality from the old Makefile and do-release.sh. It introduces automatic variables offered by Make and tries place the prerequisites so that the dependencies are in the right place. Again, use this with care, if at all. Don't use this in a production environment, because it might break things accidentially. It was fun to write, and it would be great, if it is fun to use. If you get stuck with errors, but want to push through, drop me a line and I might see, if I could help.