remove unnecessary abstractions
the original idea was to have an abstract iterator type, get all tasks
from it, and then process them in the main, but the processing itself
turned out to depend on the exact plugin.
So now we simply move all work to the plugin and have main defer to the
plugin.