feat: added a way to extend dstask externally
This was inspired by the way git allows you to add sub-commands using
the `git-command` convention and run the command using `git command`.
This is a very primitive version of that to allow users to extend dstask
with the tools they're familiar with.

It'll also allow us to develop tools (gantt charts anybody) without
cluttering up the core application.